Split out bulk mint contribution (#242)
* Split out bulk mint contribution into hook
